Move PortPlacementArgs to CreatePort.h
