+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -182,34 +182,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct PortPlacementArgs
-    {
-        static constexpr auto command = GameCommand::createPort;
-
-        PortPlacementArgs() = default;
-        explicit PortPlacementArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-            , rotation(regs.bh)
-            , type(regs.dl)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-        uint8_t type;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.di = pos.z;
-            regs.bh = rotation;
-            regs.dl = type;
-            return regs;
-        }
-    };
-
     struct PortRemovalArgs
     {
         static constexpr auto command = GameCommand::removePort;
